Fix date view obscuring reactions
There is already logic in topPadding to add some space to the top of
the view. However, this only happens if positionInMessagesSection.isTop
is false, as otherwise 0 is immediately returned. This causes the date
view to overlay on top of the reaction view, obscuring the top of the
bubbles.

Therefore, add the same bubble offset logic to the early return.