Add link preview metadata cache
This cache stores the metadata for link previews in memory. It has
performance benefits in two situations:

When scrolling past a MessageView, the metadata can be retrieved from
cache instead of regenerated (preventing the link preview from
flickering on scroll).

A ReplyMessage containing a link preview can instantly display the
cached preview, that was generated by the original Message.