Add CachedLinkPillView
This view abstracts away managing the placeholder and enriched
metadata, and swapping out the link preview views when needed.

It also transparently manages the link preview metadata cache, allowing
the consumer to simply pass a URL and not have to know whether there's
already a corresponding entry in the cache.