Add convenience urls property to AttributedString
This property extracts all URLs from the given AttributedString.

It relies on the AttributedString's link attribute, instead of
inferring link position from the text, to allow for custom styling
(where consumers of the library annotate where the links should be
themselves, via the AttributedString).

This property is useful when generating link previews, and prevents us
needing to create a new property in Message itself.