Mark ChatType and ReplyMode as Sendable
This is needed for the @Test annotation in WrappingMessagesTest, which
requires that arguments conform to the trait.

The enum was not inferred as Sendable by swift since it is publicly
available. In these cases, swift requires you to mark it as Sendable
explicitly so consumers of the library do not get confused.