Remove unused ChatError enum
This was added in ef7afea and used by the ChatService. However, in
f5461c5, the ChatService was replaced by MockChatData and no more
usages of ChatError remained. Therefore, remove this unused enum.

I found this was unused when enabling the "existential any" flag.
Instaed of migrating the enum, remove it instead.