Add LinkPillView
This view abstracts away managing the placeholder and enriched
metadata, and swapping out the link preview views when needed.

It also transparently manages the link preview metadata cache, allowing
the consumer to simply pass a URL and not have to know whether there's
already a corresponding entry in the cache.

Note that we need to disable strict concurrency checking for the
LinkPresentation framework. This is because LPLinkMetadata is not
Sendable and startFetchingMetadata(for:) uses a background thread and
nonisolated context in order to generate the metadata.

As far as I can tell, there is no way to use LinkPresentation with
strict concurrency checking, as the preview view requires a metadata
instance, and this can't cross isolation boundaries. So, the
alternative would be to write our own version of the LinkPresentation
framework that supports strict concurrency checks, which seems to be
too heavy a solution.