Replace custom date formatter with format style
I would like to have an easy and concise method to parse dates, as this
is useful in both tests and previews. While it is only used in the
WrappingMessagesTest at the moment, it will be useful in other contexts
later on.

Therefore, replace the fileprivate extension with a more standard style
formatter, which is available throughout the codebase.

This also means we can avoid using a force-unwrap in a function that
could be called from prod code, passing responsibility to unwrap the
result to the caller (which, if it is a test, can happily decide to
force-unwrap).