Align our messages with input view
The trailing padding of messages from us is made up of
horizontalStatusPadding (applied to the trailing edge of the checkmark)
and horizontalNoAvatarPadding (applied to the trailing edge of the
message view itself).

Previously, both of these were 8 points, totalling 16. Meanwhile, the
trailing padding of the input view was 12 points. This caused them to
be misaligned.

Therefore, slightly reduce both the horizontalStatusPadding and
horizontalNoAvatarPadding to 6 so that, when combined, they total 12,
and align with the edge of the input view.