Align avatar with input view
Previously, the avatar view had padding of 8 on both sides. However,
this caused the leading edge to be misaligned with the leading edge of
the input view, since the input view has horizontal padding of 12 on
both sides.

Making the avatar view padding 12 on both leading and trailing edges
looks wrong, as the gap to the message view becomes too big. Therefore,
apply a higher padding of 12 on the leading edge only.