Replace message status icons with SF symbols
The custom icons had some in-built padding which made it difficult to
align the edge of the view.

This also lets us more easily add symbols for new message states in the
future. For instance, if we want to add separate symbols for
"readByAll" or "delivered", then we can use the built-in SF symbols,
and they will look consistent with the icons for other message states.