Simplify single-sectioned WrappingMessagesTests
In many cases in WrappingMessagesTest, we are only dealing with a
single section in our test. In these cases, we use a lot of optional
chaining, which makes things slightly more complicated to understand.

Therefore, use the #require macro anytime we have a single section and
we are getting it with .first.

We have not updated usages where there are multiple sections, or where
we use optional chaining to get comment positions etc., as use of the
macro doesn't come as fluently in those situations.